moved package Projectile from Actors into demo
