+++ src/main/java/com/example/demo/Projectiles/EnemyProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo.Actors.Projectiles;
+package com.example.demo.Projectiles;
 
 public class EnemyProjectile extends Projectile {
