+++ src/main/java/com/example/demo/Projectiles/UserProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo.Actors.Projectiles;
+package com.example.demo.Projectiles;
 
 public class UserProjectile extends Projectile {
